+++ lib/ftpd/connection_throttle.rb
@@ -4,15 +4,16 @@ module Ftpd
 
   class ConnectionThrottle
 
-    DEFAULT_MAX_CONNECTIONS = 200
-    DEFAULT_MAX_CONNECTIONS_PER_IP = 5
+    DEFAULT_MAX_CONNECTIONS = nil
+    DEFAULT_MAX_CONNECTIONS_PER_IP = nil
 
-    # The maximum number of connections.
+    # The maximum number of connections, or nil if there is no limit.
     # @return [Integer]
 
     attr_accessor :max_connections
 
-    # The maximum number of connections for an IP.
+    # The maximum number of connections for an IP, or nil if there is
+    # no limit.
     # @return [Integer]
 
     attr_accessor :max_connections_per_ip
@@ -28,14 +29,26 @@ module Ftpd
     # @return [Boolean] true if the connection should be allowed
 
     def allow?(socket)
-      @connection_tracker.connections < @max_connections &&
-        @connection_tracker.connections_for(socket) < @max_connections_per_ip
+      allow_by_total_count &&
+        allow_by_ip_count(socket)
     end
 
     # Reject a connection
 
     def deny(socket)
       socket.write "421 Too many connections\r\n"
+    end
+
+    private
+
+    def allow_by_total_count
+      return true unless @max_connections
+      @connection_tracker.connections < @max_connections
+    end
+
+    def allow_by_ip_count(socket)
+      return true unless @max_connections_per_ip
+      @connection_tracker.connections_for(socket) < @max_connections_per_ip
     end
 
   end